swarm: Rename swarm test file
